Add a migrated rss.xml
Trim the fat on the old
rss feed and limit the number of
items to 10 this time.

Add an 11th apologetic
item for possibly destroying
clients when changing the <atom:id>
from blogger to my own.

New items will use the link
as the guid. Old will
have the blogger stuff
until they drop off
from new updates.

Update posts with
the rssGuid and
fix a date on
one of the posts.

Test plan:
* Save the old iamjonas.me/rss.xml file
  locally. Run
  meld <(grep 'tag:blogger' old.xml) <(grep 'tag:blogger' _site/rss.xml).
  Verify the guid for the old and new are the same
  (rss-readers will not update (I hope)).
* Do the same as above but grep for pubDate, verify
  format is the same for old and new.